feat(popup): unsaved views changes
