+++ packages/popup/src/components/ThemeEditor/index.ts
@@ -19,4 +19,4 @@ export function saveListener(event: KeyboardEvent, handleSave: () => void) {
 
 export { default as ComponentsInput } from "./ComponentsInput";
 export { default as StyleInput } from "./StyleInput";
-export { default as ViewInput } from "./ViewInput";
+export { default as ViewsInput } from "./ViewsInput";
